Allied players affected by pulsing.  [DO NOT ACCEPT YET]
This is a balancing change and should only be accepted when the balancing team have come to a decision concerning it.

Previously, gravity pulsing only affected unallied players.  In this commit, it affects everyone. (Because why wouldn't it?) 
I suppose you could argue that these pulses are actually precise beams which target enemy ships?

This commit encourages a fleet composition of smaller faster ships which pursue a larger target as opposed to just reducing the larger target's cruising speed.